refactor(ast): extract rendered-text builders into codegen.ts (#3572 1b)
Move buildJSDoc, buildObject, buildList, and objectKey out of utils/index.ts
into a dedicated utils/codegen.ts module (with indentLines kept internal), and
their tests into codegen.test.ts. utils/index.ts re-exports the same names.
